Skip flaky test
